Move App.js to component.
