+++ src/components/App.js
@@ -1,12 +1,11 @@
 import React, { Component } from 'react';
+import Menu from './menu/Menu'
 
 class App extends Component {
   render() {
     return (
       <div className="App">
-        <header className="App-header">
-
-        </header>
+        <Menu />
       </div>
     );
   }
